+++ centralia/render/__init__.py
@@ -1,7 +1,8 @@
 from .casebody import render_casebody
 from .clview import render_cl
 from .html import (opinion_text, render_body, render_headmatter, render_html,
-                   render_opinion)
+                   render_opinion, render_opinion_ingest)
 
 __all__ = ["render_casebody", "render_html", "render_opinion",
+           "render_opinion_ingest",
            "render_body", "render_headmatter", "opinion_text", "render_cl"]